requirements.txt
beautifulsoup4==4.11.1
Django==3.2.5
djangorestframework==3.13.1
lxml==4.9.1